third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/open-in-body-onload-sync-to-invalid-cross-origin-response-handling.html

<!doctype html>
<script src="/js-test-resources/js-test.js"></script>
<body onload="openXHR();">
<!-- This embed is necessary to cause the synchronous invocation of onload -->
<embed type="text/html; charset=utf-8">
<script>
window.jsTestIsAsync = true;
description('XMLHttpRequest doesn\'t crash even when open() is invoked ' +
            'synchronously to handling of a response to a cross-origin ' +
            'request.');

var xhr = new XMLHttpRequest;
var count = 10;  // The crash doesn't always happen. Repeat to capture it.

function openXHR() {
  xhr.open('GET', '/');
  openAndSendCrossOriginSimpleXHR();
}

function openAndSendCrossOriginSimpleXHR() {
  xhr.open("GET", "http://localhost:8000/xmlhttprequest/resources/test.html");
  xhr.send();
}

xhr.onerror = function() {
  --count;
  if (count <= 0) {
    setTimeout(finishJSTest, 0);
  } else {
    openAndSendCrossOriginSimpleXHR();
  }
}

openAndSendCrossOriginSimpleXHR();
</script>
